fix: 제목, 내용 표시 버그 수정
